ASoC: codecs: rv1106_codec: support rv1103b acodec
Add control 'Audio Mux Select' for rv1103b codec

Normally, we may need the acodec and SAI controller to
play and output at the same time. In this case, we need
to manually connect BCLK/LRCLK/SDO to IO when SAI function
is turned on. And, for recording, only one input can be
selected as mux that is from acodec or SAI SDI.

There is no right channel ADC on rv1103b, so the control
list result needs to remove the 'ADC Right' parts.

And the rv1106/rv1103 codec keeping the original method.

Change-Id: I637894c2e8194b99379ce2cd70ccbd9f2ec46fbc
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>