net: phy: phylink: Fixes memory leak during bind/unbind MAC driver
During each MAC drivers bind/unbind process, the kzalloc-1k
memory count increased. Because the refcount is increased
during the of_phy_find_device process, there is no decrease,
which makes it impossible to enter the release of the phy device
to release the dev. The solution is to force a put_device to be
done after the of_phy_find_device is successful.

Signed-off-by: David Wu <david.wu@rock-chips.com>
Change-Id: I6b487924b8f53b1f3c8565801105fcd46c124a67