spi: spi-rockchip-sfc: Wait for thunder boot DMA status change
The controller driver thread A and thunder-boot driver thread B are
both waiting for the pre-load idle at the same time.When thread A
obtains a valid state and continues with a new data interaction, it
will cause thread B to miss the pre-load idle signal and continue
waiting for the new data interaction to end.
So, adding a handshake signal to clear the DMA-INT status is
necessary to ensure that the thunder-boot thread receives valid
information in a timely manner.

Change-Id: I0285521d5296d357ce753333b3601484448c53d7
Signed-off-by: Jon Lin <jon.lin@rock-chips.com>